+++ src/Plate/EmptyWell.tsx
@@ -4,16 +4,26 @@ import React from 'react';
 import { PALETTE } from '../theme';
 
 import { PLATE_FLOW } from './constants';
+import { CoordinateSystem } from './types';
 import { columnForPosition, rowForPosition } from './utils';
 import { GENERAL_WELL_STYLE } from './wellUtils';
 
-export function EmptyWell(props: { position: number }) {
+export function EmptyWell<TCoordinateSystem extends CoordinateSystem>({
+  coordinateSystem,
+  position,
+}: {
+  coordinateSystem: TCoordinateSystem;
+  position: number;
+}) {
+  const row = rowForPosition(position, PLATE_FLOW, coordinateSystem);
+  const column = columnForPosition(position, PLATE_FLOW, coordinateSystem);
+
   const { setNodeRef, isOver } = useDroppable({
-    id: props.position,
+    id: position,
     data: {
       coordinates: {
-        row: rowForPosition(props.position, PLATE_FLOW),
-        column: columnForPosition(props.position, PLATE_FLOW),
+        row,
+        column,
       },
     },
   });
@@ -30,10 +40,7 @@ export function EmptyWell(props: { position: number }) {
         alignItems: 'center',
       }}
     >
-      <small>
-        {rowForPosition(props.position, PLATE_FLOW) +
-          columnForPosition(props.position, PLATE_FLOW)}
-      </small>
+      <small>{row + column}</small>
     </div>
   );
 }