+++ src/Plate/constants.ts
@@ -1,20 +1,4 @@
-import { range } from 'lodash';
+import { FlowDirection } from './types';
 
-import { Coordinates, FlowDirection } from './types';
-
-const TUBE_COUNT = 96;
-export const WELLS = range(1, TUBE_COUNT + 1);
-export const COORDINATES_COLUMNS = [
-  1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12,
-] as const satisfies ReadonlyArray<Coordinates['column']>;
-export const COORDINATES_ROWS = [
-  'A',
-  'B',
-  'C',
-  'D',
-  'E',
-  'F',
-  'G',
-  'H',
-] as const satisfies ReadonlyArray<Coordinates['row']>;
+/** Used internally when rendering the Plate component. */
 export const PLATE_FLOW = 'row' as const satisfies FlowDirection;