+++ src/Plate/index.stories.tsx
@@ -4,36 +4,51 @@ import React from 'react';
 
 import { PALETTE } from '../theme';
 
-import { COORDINATES_COLUMNS, COORDINATES_ROWS, WELLS } from './constants';
+import {
+  COORDINATE_SYSTEM_12X8,
+  CoordinateSystem12x8,
+} from './coordinateSystem12x8';
 import { PlateProps, PlateWell } from './types';
-import { coordinatesForPosition } from './utils';
+import { allCoordinateSystemPositions, coordinatesForPosition } from './utils';
 
 import { Plate } from './index';
 
 export default {
   title: 'Plate',
-  argTypes: {
-    loading: { control: 'boolean' },
-    isDraggable: { control: 'boolean' },
+  args: {
+    isDraggable: false,
+    loading: false,
   },
 };
 
-const data: Array<PlateWell> = [
+const data: Array<PlateWell<CoordinateSystem12x8>> = [
   {
-    coordinates: { row: COORDINATES_ROWS[0], column: COORDINATES_COLUMNS[6] },
+    coordinates: {
+      row: COORDINATE_SYSTEM_12X8.rows[0],
+      column: COORDINATE_SYSTEM_12X8.columns[6],
+    },
     content: <i>It renders any ReactNode</i>,
   },
   {
-    coordinates: { row: COORDINATES_ROWS[0], column: COORDINATES_COLUMNS[7] },
+    coordinates: {
+      row: COORDINATE_SYSTEM_12X8.rows[0],
+      column: COORDINATE_SYSTEM_12X8.columns[7],
+    },
     content: 'Test',
     color: PALETTE.red,
   },
   {
-    coordinates: { row: COORDINATES_ROWS[1], column: COORDINATES_COLUMNS[2] },
+    coordinates: {
+      row: COORDINATE_SYSTEM_12X8.rows[1],
+      column: COORDINATE_SYSTEM_12X8.columns[2],
+    },
     content: 'Some text',
   },
   {
-    coordinates: { row: COORDINATES_ROWS[2], column: COORDINATES_COLUMNS[2] },
+    coordinates: {
+      row: COORDINATE_SYSTEM_12X8.rows[2],
+      column: COORDINATE_SYSTEM_12X8.columns[2],
+    },
     content: (
       <>
         <p>Kontrolle</p>
@@ -45,27 +60,35 @@ const data: Array<PlateWell> = [
   },
 ];
 
-const rowFlowData: Array<PlateWell> = WELLS.map((well) => ({
-  coordinates: coordinatesForPosition(well, 'row'),
-  content: well,
-}));
+const COORDINATE_SYSTEM_96_WELL_POSITIONS = allCoordinateSystemPositions(
+  COORDINATE_SYSTEM_12X8,
+);
 
-const columnFlowData: Array<PlateWell> = WELLS.map((well) => ({
-  coordinates: coordinatesForPosition(well, 'column'),
-  content: well,
-}));
+const rowFlowData: Array<PlateWell<CoordinateSystem12x8>> =
+  COORDINATE_SYSTEM_96_WELL_POSITIONS.map((well) => ({
+    coordinates: coordinatesForPosition(well, 'row', COORDINATE_SYSTEM_12X8),
+    content: well,
+  }));
 
-const Template: Story<Partial<PlateProps>> = function Template(args) {
-  return (
-    <Plate
-      data={null}
-      dndContextProps={{
-        onDragEnd: action('onDragEnd'), // dataLocation: `const sourceData = e.active.data.current; const targetData = e.over?.data.current;`
-      }}
-      {...args}
-    />
-  );
-};
+const columnFlowData: Array<PlateWell<CoordinateSystem12x8>> =
+  COORDINATE_SYSTEM_96_WELL_POSITIONS.map((well) => ({
+    coordinates: coordinatesForPosition(well, 'column', COORDINATE_SYSTEM_12X8),
+    content: well,
+  }));
+
+const Template: Story<Partial<PlateProps<CoordinateSystem12x8>>> =
+  function Template(args) {
+    return (
+      <Plate
+        data={null}
+        coordinateSystem={COORDINATE_SYSTEM_12X8}
+        dndContextProps={{
+          onDragEnd: action('onDragEnd'), // dataLocation: `const sourceData = e.active.data.current; const targetData = e.over?.data.current;`
+        }}
+        {...args}
+      />
+    );
+  };
 
 export const Default = Template.bind({});
 Default.args = {