+++ src/Select/formInput.ts
@@ -10,8 +10,17 @@ export function objectToFormInputOptions<
   TLabels extends Record<string, TLabel>,
   TLabel extends ReactNode,
 >(labels: TLabels): Array<FormInputOption<keyof TLabels, TLabel>> {
-  return Object.entries(labels).map(([key, value]) => ({
-    value: key as keyof TLabels,
+  return objectEntriesToFormInputOptions(Object.entries(labels));
+}
+
+export function objectEntriesToFormInputOptions<
+  TLabelEntries extends Array<[string, TLabel]>,
+  TLabel extends ReactNode,
+>(
+  labelEntries: TLabelEntries,
+): Array<FormInputOption<TLabelEntries[number][0], TLabel>> {
+  return labelEntries.map(([key, value]) => ({
+    value: key as TLabelEntries[number][0],
     label: value,
   }));
 }